refactor: sectionAddManager 책임 sections로 이동
