+++ src/main/java/wooteco/subway/service/SubwayService.java
@@ -9,7 +9,6 @@ import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Line;
 import wooteco.subway.domain.Lines;
 import wooteco.subway.domain.Section;
-import wooteco.subway.domain.SectionAddManager;
 import wooteco.subway.domain.Sections;
 import wooteco.subway.domain.Station;
 import wooteco.subway.domain.Stations;
@@ -126,9 +125,9 @@ public class SubwayService {
     }
 
     public void addSection(Long lineId, SectionRequest sectionRequest) {
-        SectionAddManager addManager = SectionAddManager.of(sectionDao.findByLineId(lineId));
-        List<Section> updatedSections = addManager.add(sectionRequest.toEntity(lineId));
+        Sections sections = new Sections(sectionDao.findByLineId(lineId));
+        sections.add(sectionRequest.toEntity(lineId));
         sectionDao.delete(lineId);
-        updatedSections.forEach(sectionDao::save);
+        sections.forEach(sectionDao::save);
     }
 }
